protein-folder/requirements.txt
torch
numpy
matplotlib
scikit-learn
biopython
requests
